Better sourcing
Now will attempt to inteligently source dependencies. Checks if it's in
a ./lib/ directory, else if it's available in your $PATH. If a
dependency has already been sourced (in the case of multiple files all
requiring the same dependency) it will only source once. Ensures that
any setup logic isn't run repeatedly.